API references

HetaSimulator.heta_build — Method

heta_build(  
  heta_index::AbstractString;
  declaration::String = "platform",
  skip_export::Bool = false,
  log_mode::String = "error",
  debug::Bool = false,
  julia_only::Bool = false,
  dist_dir::String = "dist",
  meta_dir::String = "meta",
  source::String = "index.heta",
  type::String = "heta"
)

Builds the model from Heta-based reactions

HetaSimulator.heta_update — Method

heta_update(version::String)

Installs heta-compiler from NPM.

HetaSimulator.load_jlplatform — Method

load_jlplatform(  
  model_jl::AbstractString; 
  rm_out::Bool = false
)

Loads prebuild julia model as part of platform

HetaSimulator.load_platform — Method

load_platform(  
  heta_index::AbstractString;
  rm_out::Bool = true,
  julia_only::Bool = true, 
  dist_dir::String = ".",
  source::String = "index.heta",
  type::String = "heta",
  kwargs...
)

Converts heta model to Julia and outputs platform type

HetaSimulator.sim — Method

sim(model::Model; 
  parameters::Vector{Pair{Symbol,Float64}} = Pair{Symbol,Float64}[],
  measurements::Vector{AbstractMeasurementPoint} = AbstractMeasurementPoint[],
  events_active::Union{Nothing, Vector{Pair{Symbol,Bool}}} = Pair{Symbol,Bool}[],
  events_save::Union{Tuple,Vector{Pair{Symbol, Tuple{Bool, Bool}}}}=(true,true), 
  observables::Union{Nothing,Vector{Symbol}} = nothing,
  saveat::Union{Nothing,AbstractVector} = nothing,
  tspan::Union{Nothing,Tuple} = nothing,
  save_scope::Bool=true,
  time_type::DataType=Float64,
  kwargs...)

Simulate model of type Model. Returns SimResults type.

Arguments:

model : model of type Model

parameters : Vector of Pairs containing constants' names and values. Overwrites default model's values. Default is empty vector

measurements : Vector of measurements. Default is empty vector

events_active : Vector of Pairs containing events' names and true/false values. Overwrites default model's values. Default is empty vector

events_save : Tuple or Vector{Tuple} marking whether to save solution before and after event. Default is (true,true) for all events

observables : names of output observables. Overwrites default model's values. Default is empty vector

saveat : time points, where solution should be saved. Default nothing values stands for saving solution at timepoints reached by the solver

tspan : time span for the ODE problem

save_scope : should scope be saved together with solution. Default is true

kwargs : other solver related arguments supported by sim(cond::Cond)

HetaSimulator.sim — Method

sim(platform::Platform; 
  conditions::Union{AbstractVector{Symbol}, Nothing} = nothing,
  kwargs...) where {C<:AbstractCond}

Simulate conditions included in platform. Returns Vector{Pair}.

Arguments:

platform : platform of Platform type

conditions : Vector containing names of conditions included in platform. Default value nothing stands for all conditions in the platform

kwargs : other kwargs supported by sim(condition_pairs::Vector{Pair})

HetaSimulator.sim — Method

sim(conditions::AbstractVector{C}; kwargs...) where {C<:AbstractCond}

Simulate multiple conditions. Returns Vector{Pair}.

Arguments:

conditions : Vector containing names and conditions of type Cond

kwargs : other kwargs supported by sim(condition_pairs::Vector{Pair})

HetaSimulator.sim — Method

sim(cond::Cond; 
  parameters_upd::Union{Nothing, Vector{P}}=nothing,
  alg=DEFAULT_ALG, 
  reltol=DEFAULT_SIMULATION_RELTOL, 
  abstol=DEFAULT_SIMULATION_ABSTOL,
  kwargs...)

Simulate single condition cond. Returns SimResults type.

Arguments:

cond : simulation condition of type Cond

parameters_upd : constants, which overwrite both Model and Cond constants. Default is nothing

alg : ODE solver. See SciML docs for details. Default is AutoTsit5(Rosenbrock23())

reltol : relative tolerance. Default is 1e-3

abstol : relative tolerance. Default is 1e-6

kwargs : other solver related arguments supported by DiffEqBase.solve. See SciML docs for details

HetaSimulator.sim — Method

sim(condition_pairs::Vector{P}; 
  parameters_upd::Union{Nothing, Vector}=nothing,
  alg=DEFAULT_ALG, 
  reltol=DEFAULT_SIMULATION_RELTOL, 
  abstol=DEFAULT_SIMULATION_ABSTOL,
  parallel_type=EnsembleSerial(),
  kwargs...) where P<:Pair

Simulate multiple conditions. Returns Vector{Pair}.

Arguments:

condition_pairs : vector of pairs containing names and conditions of type Cond

parameters_upd : constants, which overwrite both Model and Cond constants. Default is nothing

alg : ODE solver. See SciML docs for details. Default is AutoTsit5(Rosenbrock23())

reltol : relative tolerance. Default is 1e-3

abstol : relative tolerance. Default is 1e-6

parallel_type : type of multiple simulations parallelism. Default is no parallelism. See SciML docs for details

kwargs : other solver related arguments supported by DiffEqBase.solve. See SciML docs for details