API references

HetaSimulator.heta_build — Method

heta_build(  
  heta_index::AbstractString;
  declaration::String = "platform",
  skip_export::Bool = false,
  log_mode::String = "error",
  debug::Bool = false,
  julia_only::Bool = false,
  dist_dir::String = "dist",
  meta_dir::String = "meta",
  source::String = "index.heta",
  type::String = "heta"
)

Builds the model from Heta-based reactions

HetaSimulator.heta_update — Method

heta_update(version::String)

Installs heta-compiler from NPM.

HetaSimulator.load_jlplatform — Method

load_jlplatform(  
  model_jl::AbstractString; 
  rm_out::Bool = false
)

Loads prebuild julia model as part of platform

HetaSimulator.load_platform — Method

load_platform(  
  heta_index::AbstractString;
  rm_out::Bool = true,
  julia_only::Bool = true, 
  dist_dir::String = ".",
  source::String = "index.heta",
  type::String = "heta",
  kwargs...
)

Converts heta model to Julia and outputs platform type

HetaSimulator.sim — Method

sim(cond::Cond; 
  alg=DEFAULT_ALG, 
  reltol=DEFAULT_SIMULATION_RELTOL, 
  abstol=DEFAULT_SIMULATION_ABSTOL,
  kwargs...)

Simulate single condition cond. Returns SimResults type.

Arguments:

cond : simulation condition of type Cond

alg : ODE solver. See SciML docs for details. Default is AutoTsit5(Rosenbrock23())

reltol : relative tolerance. Default is 1e-3

abstol : relative tolerance. Default is 1e-6

kwargs : other solver related arguments supported by DiffEqBase.solve. See SciML docs for details

HetaSimulator.sim — Method

sim(cond::Cond; 
  alg=DEFAULT_ALG, 
  reltol=DEFAULT_SIMULATION_RELTOL, 
  abstol=DEFAULT_SIMULATION_ABSTOL,
  kwargs...)

Simulate single condition cond. Returns SimResults type.

Arguments:

cond : simulation condition of type Cond

alg : ODE solver. See SciML docs for details. Default is AutoTsit5(Rosenbrock23())

reltol : relative tolerance. Default is 1e-3

abstol : relative tolerance. Default is 1e-6

kwargs : other solver related arguments supported by DiffEqBase.solve. See SciML docs for details

HetaSimulator.sim — Method

sim(condition_pairs::Vector{P}; 
  alg=DEFAULT_ALG, 
  reltol=DEFAULT_SIMULATION_RELTOL, 
  abstol=DEFAULT_SIMULATION_ABSTOL,
  parallel_type=EnsembleSerial(),
  kwargs...) where P<:Pair

Simulate multiple conditions. Returns Vector{Pair}.

Arguments:

condition_pairs : vector of pairs containing names and conditions of type Cond

alg : ODE solver. See SciML docs for details. Default is AutoTsit5(Rosenbrock23())

reltol : relative tolerance. Default is 1e-3

abstol : relative tolerance. Default is 1e-6

parallel_type : type of multiple simulations parallelism. Default is no parallelism. See SciML docs for details

kwargs : other solver related arguments supported by DiffEqBase.solve. See SciML docs for details