API references

HetaSimulator.Condition — Method

Condition(model::Model;
  measurements::Vector{AbstractMeasurementPoint}=AbstractMeasurementPoint[],
  observables::Union{Nothing,Vector{Symbol}}=nothing,
  parameters::Vector{Pair{Symbol,Float64}} = Pair{Symbol,Float64}[],
  events_active::Union{Nothing, Vector{Pair{Symbol,Bool}}} = Pair{Symbol,Bool}[],
  events_save::Union{Tuple,Vector{Pair{Symbol, Tuple{Bool, Bool}}}} = (true,true), 
  saveat::Union{Nothing,AbstractVector} = nothing,
  tspan::Union{Nothing,Tuple} = nothing,
  save_scope::Bool = true,
)

Builds simulation condition of type Condition Example: Condition(model; tspan = (0., 200.), saveat = [0.0, 150., 250.])

Arguments:

model : model of type Model

measurements : Vector of measurements. Default is empty Vector{AbstractMeasurementPoint}

observables : names of output observables. Overwrites default model's values. Default is nothing

parameters : parameters variation setup. Default is empty Vector{Pair}

events_active : Vector of Pairs containing events' names and true/false values. Overwrites default model's values. Default is empty Vector{Pair}

events_save : Tuple or Vector{Tuple} marking whether to save solution before and after event. Default is (true,true) for all events

saveat : time points, where solution should be saved. Default nothing values stands for saving solution at timepoints reached by the solver

tspan : time span for the ODE problem

save_scope : should scope be saved together with solution. Default is true

HetaSimulator.fit — Method

fit(platform::Platform,
  params::Vector{Pair{Symbol,Float64}};
  conditions::Union{AbstractVector{Symbol}, Nothing} = nothing,
  kwargs...
) where C<:AbstractCond

Fit parameters to experimental measurements. Returns FitResults type. Example: fit(platform, [:k1=>0.1,:k2=>0.2,:k3=>0.3];conditions=[:cond2,:cond3])

Arguments:

platform : platform of Platform type

params : optimization parameters and their initial values

conditions : vector of conditions of type Condition or nothing to fit all conditions. Default is nothing

kwargs : other solver related arguments supported by fit(condition_pairs::Vector{<:Pair}, params::Vector{<:Pair}

HetaSimulator.fit — Method

fit(condition_pairs::AbstractVector{Pair{Symbol, C}},
  params::Vector{Pair{Symbol,Float64}};
  alg=DEFAULT_ALG,
  reltol=DEFAULT_FITTING_RELTOL,
  abstol=DEFAULT_FITTING_ABSTOL,
  parallel_type=EnsembleSerial(),
  ftol_abs = 0.0,
  ftol_rel = 1e-4, 
  xtol_rel = 0.0,
  xtol_abs = 0.0, 
  fit_alg = :LN_NELDERMEAD,
  maxeval = 10000,
  maxtime = 0.0,
  lbounds = fill(0.0, length(params)),
  ubounds = fill(Inf, length(params)),
  kwargs... 
) where C<:AbstractCond

Fit parameters to experimental measurements. Returns FitResults type. Example: fit([:x=>cond2, :y=>cond3, :z=>cond4], [:k1=>0.1,:k2=>0.2,:k3=>0.3])

Arguments:

condition_pairs : vector of pairs containing names and conditions of type Condition

params : optimization parameters and their initial values

alg : ODE solver. See SciML docs for details. Default is AutoTsit5(Rosenbrock23())

reltol : relative tolerance. Default is 1e-6

abstol : relative tolerance. Default is 1e-8

parallel_type : parallel setup. See SciML docs for details. Default is no parallelism: EnsembleSerial()

ftol_abs : absolute tolerance on function value. See NLopt.jl docs for details. Default is 0.0

ftol_rel : relative tolerance on function value. See NLopt.jl docs for details. Default is 1e-4

xtol_rel : relative tolerance on optimization parameters. See NLopt.jl docs for details. Default is 0.0

xtol_rel : absolute tolerance on optimization parameters. See NLopt.jl docs for details. Default is 0.0

fit_alg : fitting algorithm. See NLopt.jl docs for details. Default is :LN_NELDERMEAD

maxeval : maximum number of function evaluations. See NLopt.jl docs for details. Default is 1e4

maxtime : maximum optimization time (in seconds). See NLopt.jl docs for details. Default is 0

lbounds : lower parameters bounds. See NLopt.jl docs for details. Default is fill(0.0, length(params))

ubounds : upper parameters bounds. See NLopt.jl docs for details. Default is fill(Inf, length(params))

kwargs : other solver related arguments supported by DiffEqBase.solve. See SciML docs for details

HetaSimulator.fit — Method

fit(conditions::AbstractVector{C},
  params::Vector{Pair{Symbol,Float64}};
  kwargs...
) where C<:AbstractCond

Fit parameters to experimental measurements. Returns FitResults type. Example: fit([cond2, cond3, cond4], [:k1=>0.1,:k2=>0.2,:k3=>0.3])

Arguments:

conditions : vector of conditions of type Condition

params : optimization parameters and their initial values

kwargs : other solver related arguments supported by fit(condition_pairs::Vector{<:Pair}, params::Vector{<:Pair}

HetaSimulator.heta_build — Method

heta_build(
  heta_index::AbstractString;
  declaration::String = "platform",
  skip_export::Bool = false,
  log_mode::String = "error",
  debug::Bool = false,
  julia_only::Bool = false,
  dist_dir::String = "dist",
  meta_dir::String = "meta",
  source::String = "index.heta",
  type::String = "heta"
)

Builds the model from Heta-based reactions

HetaSimulator.heta_update — Method

heta_update(version::String)

Installs heta-compiler from NPM.

HetaSimulator.load_jlplatform — Method

load_jlplatform(  
  model_jl::AbstractString; 
  rm_out::Bool = false
)

Loads prebuild julia model as part of platform

HetaSimulator.load_platform — Method

load_platform(  
  heta_index::AbstractString;
  rm_out::Bool = true,
  julia_only::Bool = true, 
  dist_dir::String = ".",
  source::String = "index.heta",
  type::String = "heta",
  kwargs...
)

Converts heta model to Julia and outputs platform type

HetaSimulator.mc — Method

mc(cond::Condition,
  params::DataFrame,
  num_iter::Int64;
  kwargs...
)

Run Monte-Carlo simulations with single condition cond. Returns MCResults type. Example: mc(cond1, DataFrame(k2=rand(3),k3=rand(3)), 1000)

Arguments:

cond : simulation condition of type Condition

params : DataFrame with pre-generated parameters.

num_iter : number of Monte-Carlo iterations

kwargs : other solver related arguments supported by mc(cond::Condition, params::Vector, num_iter::Int64)

HetaSimulator.mc — Method

mc(cond_pairs::Vector{<:Pair},
  params::Vector{<:Pair},
  num_iter::Int64;
  verbose=false,
  alg=DEFAULT_ALG,
  reltol=DEFAULT_SIMULATION_RELTOL,
  abstol=DEFAULT_SIMULATION_ABSTOL,
  parallel_type=EnsembleSerial(),
  kwargs...
)

Run Monte-Carlo simulations with single condition cond. Returns Vector{MCResults} type. Example: mc([:c1=>cond1,:c2=>cond2], [:k2=>Normal(1e-3,1e-4), :k3=>Uniform(1e-4,1e-2)], 1000)

Arguments:

cond_pairs : vector of pairs containing names and conditions of type Condition

params : parameters variation setup

num_iter : number of Monte-Carlo iterations

verbose : print iteration progress. Default is false

alg : ODE solver. See SciML docs for details. Default is AutoTsit5(Rosenbrock23())

reltol : relative tolerance. Default is 1e-3

abstol : relative tolerance. Default is 1e-6

parallel_type : parallel setup. See SciML docs for details. Default is no parallelism: EnsembleSerial()

kwargs : other solver related arguments supported by DiffEqBase.solve. See SciML docs for details

HetaSimulator.mc — Method

mc(conds::Vector{<:AbstractCond},
  params::Vector{<:Pair},
  num_iter::Int64;
  kwargs...
)

Run Monte-Carlo simulations with single condition cond. Returns Vector{MCResults} type. Example: mc([cond1,cond2], [:k2=>Normal(1e-3,1e-4), :k3=>Uniform(1e-4,1e-2)], 1000)

Arguments:

cond_pairs : vector of conditions of type Condition

params : parameters variation setup

num_iter : number of Monte-Carlo iterations

kwargs : other solver related arguments supported by mc(cond_pairs::Vector{<:Pair}, params::Vector, num_iter::Int64)

HetaSimulator.mc — Method

mc(cond::Condition,
  params::Vector{<:Pair},
  num_iter::Int64;
  verbose=false,
  alg=DEFAULT_ALG,
  reltol=DEFAULT_SIMULATION_RELTOL,
  abstol=DEFAULT_SIMULATION_ABSTOL,
  parallel_type=EnsembleSerial(),
  kwargs...
)

Run Monte-Carlo simulations with single Condition. Returns MCResults type. Example: mc(cond, [:k2=>Normal(1e-3,1e-4), :k3=>Uniform(1e-4,1e-2)], 1000)

Arguments:

cond : simulation condition of type Condition

params : parameters variation setup

num_iter : number of Monte-Carlo iterations

verbose : print iteration progress. Default is false

alg : ODE solver. See SciML docs for details. Default is AutoTsit5(Rosenbrock23())

reltol : relative tolerance. Default is 1e-3

abstol : relative tolerance. Default is 1e-6

parallel_type : parallel setup. See SciML docs for details. Default is no parallelism: EnsembleSerial()

kwargs : other solver related arguments supported by DiffEqBase.solve. See SciML docs for details

HetaSimulator.mc — Method

mc(model::Model,
  params::Vector{<:Pair},
  num_iter::Int64;
  measurements::Vector{AbstractMeasurementPoint} = AbstractMeasurementPoint[],
  events_active::Union{Nothing, Vector{Pair{Symbol,Bool}}} = Pair{Symbol,Bool}[],
  events_save::Union{Tuple,Vector{Pair{Symbol, Tuple{Bool, Bool}}}}=(true,true), 
  observables::Union{Nothing,Vector{Symbol}} = nothing,
  saveat::Union{Nothing,AbstractVector} = nothing,
  tspan::Union{Nothing,Tuple} = nothing,
  save_scope::Bool=false,
  time_type::DataType=Float64,
  kwargs...
)

Run Monte-Carlo simulations with Model. Returns MCResults type. Example: mc(model, [:k2=>Normal(1e-3,1e-4), :k3=>Uniform(1e-4,1e-2)], 1000)

Arguments:

model : model of type Model

params : parameters variation setup

num_iter : number of Monte-Carlo iterations

measurements : Vector of measurements. Default is empty vector

events_active : Vector of Pairs containing events' names and true/false values. Overwrites default model's values. Default is empty vector

events_save : Tuple or Vector{Tuple} marking whether to save solution before and after event. Default is (true,true) for all events

observables : names of output observables. Overwrites default model's values. Default is empty vector

saveat : time points, where solution should be saved. Default nothing values stands for saving solution at timepoints reached by the solver

tspan : time span for the ODE problem

save_scope : should scope be saved together with solution. Default is false

kwargs : other solver related arguments supported by mc(cond::Condition, params::Vector, num_iter::Int64)

HetaSimulator.mc — Method

mc(platform::Platform, 
  params::Vector{<:Pair},
  num_iter::Int64;
  kwargs...
)

Run Monte-Carlo simulations with single condition cond. Returns Vector{MCResults} type. Example: mc(platform, [:k2=>Normal(1e-3,1e-4), :k3=>Uniform(1e-4,1e-2)], 1000)

Arguments:

platform : platform of Platform type

params : parameters variation setup

num_iter : number of Monte-Carlo iterations

kwargs : other solver related arguments supported by mc(cond_pairs::Vector{<:Pair}, params::Vector, num_iter::Int64)

HetaSimulator.sim — Method

sim(model::Model; 
  parameters::Vector{Pair{Symbol,Float64}} = Pair{Symbol,Float64}[],
  measurements::Vector{AbstractMeasurementPoint} = AbstractMeasurementPoint[],
  events_active::Union{Nothing, Vector{Pair{Symbol,Bool}}} = Pair{Symbol,Bool}[],
  events_save::Union{Tuple,Vector{Pair{Symbol, Tuple{Bool, Bool}}}}=(true,true), 
  observables::Union{Nothing,Vector{Symbol}} = nothing,
  saveat::Union{Nothing,AbstractVector} = nothing,
  tspan::Union{Nothing,Tuple} = nothing,
  save_scope::Bool=true,
  time_type::DataType=Float64,
  kwargs...)

Simulate model of type Model. Returns SimResults type. Example: sim(model; tspan = (0., 200.), parameters_upd = [:k1=>0.01])

Arguments:

model : model of type Model

parameters : Vector of Pairs containing constants' names and values. Overwrites default model's values. Default is empty vector

measurements : Vector of measurements. Default is empty vector

events_active : Vector of Pairs containing events' names and true/false values. Overwrites default model's values. Default is empty vector

events_save : Tuple or Vector{Tuple} marking whether to save solution before and after event. Default is (true,true) for all events

observables : names of output observables. Overwrites default model's values. Default is empty vector

saveat : time points, where solution should be saved. Default nothing values stands for saving solution at timepoints reached by the solver

tspan : time span for the ODE problem

save_scope : should scope be saved together with solution. Default is true

kwargs : other solver related arguments supported by sim(cond::Condition)

HetaSimulator.sim — Method

sim(platform::Platform; 
  conditions::Union{AbstractVector{Symbol}, Nothing} = nothing,
  kwargs...) where {C<:AbstractCond}

Simulate conditions included in platform. Returns Vector{Pair}. Example: sim(platform)

Arguments:

platform : platform of Platform type

conditions : Vector containing names of conditions included in platform. Default value nothing stands for all conditions in the platform

kwargs : other kwargs supported by sim(condition_pairs::Vector{Pair})

HetaSimulator.sim — Method

sim(conditions::AbstractVector{C}; kwargs...) where {C<:AbstractCond}

Simulate multiple conditions. Returns Vector{Pair}. Example: sim([cond1, cond2, cond3])

Arguments:

conditions : Vector containing names and conditions of type Condition

kwargs : other kwargs supported by sim(condition_pairs::Vector{Pair})

HetaSimulator.sim — Method

sim(cond::Condition; 
  parameters_upd::Union{Nothing, Vector{P}}=nothing,
  alg=DEFAULT_ALG, 
  reltol=DEFAULT_SIMULATION_RELTOL, 
  abstol=DEFAULT_SIMULATION_ABSTOL,
  kwargs...)

Simulate single condition cond. Returns SimResults type. Example: Condition(model; tspan = (0., 200.), saveat = [0.0, 150., 250.]) |> sim

Arguments:

cond : simulation condition of type Condition

parameters_upd : constants, which overwrite both Model and Condition constants. Default is nothing

alg : ODE solver. See SciML docs for details. Default is AutoTsit5(Rosenbrock23())

reltol : relative tolerance. Default is 1e-3

abstol : relative tolerance. Default is 1e-6

kwargs : other solver related arguments supported by DiffEqBase.solve. See SciML docs for details

HetaSimulator.sim — Method

sim(condition_pairs::Vector{P}; 
  parameters_upd::Union{Nothing, Vector}=nothing,
  alg=DEFAULT_ALG, 
  reltol=DEFAULT_SIMULATION_RELTOL, 
  abstol=DEFAULT_SIMULATION_ABSTOL,
  parallel_type=EnsembleSerial(),
  kwargs...) where P<:Pair

Simulate multiple conditions. Returns Vector{Pair}. Example: sim([:x => cond1, :y=>cond2, :z=>cond3])

Arguments:

condition_pairs : vector of pairs containing names and conditions of type Condition

parameters_upd : constants, which overwrite both Model and Condition constants. Default is nothing

alg : ODE solver. See SciML docs for details. Default is AutoTsit5(Rosenbrock23())

reltol : relative tolerance. Default is 1e-3

abstol : relative tolerance. Default is 1e-6

parallel_type : type of multiple simulations parallelism. Default is no parallelism. See SciML docs for details

kwargs : other solver related arguments supported by DiffEqBase.solve. See SciML docs for details