API references

HetaSimulator.FitResults — Type

struct FitResults{L<:Real, I}
  obj::L
  optim::Vector{Pair{Symbol,Float64}}
  status::Symbol
  numevals::I
end

Results of fit.

Use optim method to get optimal values.

Use object method to get optimal objective function.

Use status to get status.

HetaSimulator.MCResults — Type

struct MCResults{S,C} <: AbstractResults
  sim::S
  saveat::Bool
  scenario::C
end

Structure storing results of [mc]{@ref} method applied for one Scenario.

To convert into tabular format, use DataFrame method.

The base visulaization can be done with plot method.

HetaSimulator.Model — Type

struct Model{IF,OF,EV,SG,C,EA} <: AbstractModel
  init_func::IF
  ode_func::OF
  events::EV
  saving_generator::SG
  records_output::AbstractVector{Pair{Symbol,Bool}}
  constants::C
  events_active::EA
end

Structure storing core properties of ODE model. This represent the content of one namespace from a Heta platform.

To get list of model content use methods: constants(model), records(model), switchers(model).

To get the default model options use methods: parameters(model), events_active(model), events_save(model), observables(model). These options can be rewritten by a [Scenario]{@ref}.

HetaSimulator.Platform — Type

struct Platform{M,C}
  models::Dict{Symbol,M}     # dictionary storing Models
  scenarios::Dict{Symbol,C} # dictionary storing Scenarios
end

The main storage representing a modeling platform. Typically HetaSimulator works with one platform object which can include several models and scenarios.

Usually a Platform is created based on Heta formatted files using [load_platform]{@ref}.

To get the platform content use methods: models(platform), `scenarios(platform).

HetaSimulator.Scenario — Type

struct Scenario{F,P,M} <: AbstractScenario init_func::F prob::P measurements::M end

Type representing simulation conditions, i.e. model variant with updated constants and outputs.

To get the internal properties use methods: saveat(scenario), tspan(scenario), parameters(scenario), measurements(scenario)

HetaSimulator.Scenario — Method

Scenario(model::Model;
  measurements::Vector{AbstractMeasurementPoint}=AbstractMeasurementPoint[],
  observables::Union{Nothing,Vector{Symbol}}=nothing,
  parameters::Vector{Pair{Symbol,Float64}} = Pair{Symbol,Float64}[],
  events_active::Union{Nothing, Vector{Pair{Symbol,Bool}}} = Pair{Symbol,Bool}[],
  events_save::Union{Tuple,Vector{Pair{Symbol, Tuple{Bool, Bool}}}} = (true,true), 
  saveat::Union{Nothing,AbstractVector} = nothing,
  tspan::Union{Nothing,Tuple} = nothing,
  save_scope::Bool = true,
)

Builds simulation scenario of type Scenario

Example: Scenario(model; tspan = (0., 200.), saveat = [0.0, 150., 250.])

Arguments:

model : model of type Model

measurements : Vector of measurements. Default is empty Vector{AbstractMeasurementPoint}

observables : names of output observables. Overwrites default model's values. Default is nothing

parameters : Vector of Pairs containing constants' names and values. Overwrites default model's values. Default is empty vector.

events_active : Vector of Pairs containing events' names and true/false values. Overwrites default model's values. Default is empty Vector{Pair}

events_save : Tuple or Vector{Tuple} marking whether to save solution before and after event. Default is (true,true) for all events

saveat : time points, where solution should be saved. Default nothing values stands for saving solution at timepoints reached by the solver

tspan : time span for the ODE problem

save_scope : should scope be saved together with solution. Default is true

HetaSimulator.SimResults — Type

struct SimResults{S, C<:Scenario} <: AbstractResults
  sim::S
  scenario::C 
end

Structure storing results from [sim]{@ref} method applied for one [Scenario]{@ref}.

To get the content use methods: status(results), times(results), vals(results), parameters(results), observables(results).

The results can be transformed using DataFrame method or visualized using plot method.

HetaSimulator.add_measurements! — Method

add_measurements!(
  platform::Platform,
  df::DataFrame;
  kwargs...
)

Adds measurements to Scenario

Arguments:

platform : platform of Platform type

df : DataFrame with measurements, typically obtained with read_measurements function

subset : subset of measurements which will be added to the Scenario. Default Pair{Symbol, Symbol}[] adds all measurements from the df

HetaSimulator.add_measurements! — Method

add_measurements!(
  scenario::Scenario,
  df::DataFrame;
  kwargs...
)

Adds measurements to Scenario

Arguments:

scenario : simulation scenario of type Scenario

df : DataFrame with measurements, typically obtained with read_measurements function

subset : subset of measurements which will be added to the Scenario. Default Pair{Symbol, Symbol}[] adds all measurements from the df

HetaSimulator.add_scenarios! — Method

add_scenarios!(
  platform::Platform,
  df::DataFrame;
  subset::AbstractVector{P} = Pair{Symbol, Symbol}[]
) where P <: Pair{Symbol, Symbol}

Adds a new Scenario to the Platform

Arguments:

platform : platform of Platform type

df : DataFrame with scenarios setup, typically obtained with read_scenarios function

subset : subset of scenarios which will be added to the platform. Default Pair{Symbol, Symbol}[] adds all scenarios from the df

HetaSimulator.fit — Method

fit(platform::Platform,
  params;
  scenarios::Union{AbstractVector{Symbol}, Nothing} = nothing,
  kwargs...
) where C<:AbstractScenario

Fit parameters to experimental measurements. Returns FitResults type.

Example: fit(platform, [:k1=>0.1,:k2=>0.2,:k3=>0.3];scenarios=[:scn2,:scn3])

Arguments:

platform : platform of Platform type

params : optimization parameters and their initial values

scenarios : vector of scenarios of type Scenario or nothing to fit all scenarios. Default is nothing

kwargs : other solver related arguments supported by fit(scenario_pairs::Vector{<:Pair}, params::Vector{<:Pair}

HetaSimulator.fit — Method

fit(scenario_pairs::AbstractVector{Pair{Symbol, C}},
  params_df::DataFrame;
  kwargs...
) where C<:AbstractScenario

Fit parameters to experimental measurements. Returns FitResults type.

Arguments:

scenario_pairs : vector of pairs containing names and scenarios of type Scenario

params : DataFrame with optimization parameters setup and their initial values

kwargs : other solver related arguments supported by fit(scenario_pairs::Vector{<:Pair}, params::Vector{<:Pair}

HetaSimulator.fit — Method

fit(scenarios::AbstractVector{C},
  params;
  kwargs...
) where C<:AbstractScenario

Fit parameters to experimental measurements. Returns FitResults type.

Example: fit([scn2, scn3, scn4], [:k1=>0.1,:k2=>0.2,:k3=>0.3])

Arguments:

scenarios : vector of scenarios of type Scenario

params : optimization parameters and their initial values

kwargs : other solver related arguments supported by fit(scenario_pairs::Vector{<:Pair}, params::Vector{<:Pair}

HetaSimulator.fit — Method

fit(scenario_pairs::AbstractVector{Pair{Symbol, C}},
  params::Vector{Pair{Symbol,Float64}};
  alg=DEFAULT_ALG,
  reltol=DEFAULT_FITTING_RELTOL,
  abstol=DEFAULT_FITTING_ABSTOL,
  parallel_type=EnsembleSerial(),
  ftol_abs = 0.0,
  ftol_rel = 1e-4, 
  xtol_rel = 0.0,
  xtol_abs = 0.0, 
  fit_alg = :LN_NELDERMEAD,
  maxeval = 10000,
  maxtime = 0.0,
  lbounds = fill(0.0, length(params)),
  ubounds = fill(Inf, length(params)),
  kwargs... 
) where C<:AbstractScenario

Fit parameters to experimental measurements. Returns FitResults type.

Example: fit([:x=>scn2, :y=>scn3, :z=>scn4], [:k1=>0.1,:k2=>0.2,:k3=>0.3])

Arguments:

scenario_pairs : vector of pairs containing names and scenarios of type Scenario

params : optimization parameters and their initial values

parameters_upd : constants, which overwrite both Model and Scenario constants. Default is nothing

alg : ODE solver. See SciML docs for details. Default is AutoTsit5(Rosenbrock23())

reltol : relative tolerance. Default is 1e-6

abstol : relative tolerance. Default is 1e-8

parallel_type : parallel setup. See SciML docs for details. Default is no parallelism: EnsembleSerial()

ftol_abs : absolute tolerance on function value. See NLopt.jl docs for details. Default is 0.0

ftol_rel : relative tolerance on function value. See NLopt.jl docs for details. Default is 1e-4

xtol_rel : relative tolerance on optimization parameters. See NLopt.jl docs for details. Default is 0.0

xtol_rel : absolute tolerance on optimization parameters. See NLopt.jl docs for details. Default is 0.0

fit_alg : fitting algorithm. See NLopt.jl docs for details. Default is :LN_NELDERMEAD

maxeval : maximum number of function evaluations. See NLopt.jl docs for details. Default is 1e4

maxtime : maximum optimization time (in seconds). See NLopt.jl docs for details. Default is 0

lbounds : lower parameters bounds. See NLopt.jl docs for details. Default is fill(0.0, length(params))

ubounds : upper parameters bounds. See NLopt.jl docs for details. Default is fill(Inf, length(params))

scale : scale of the parameters (supports :lin, :log, :log10) to be used during fitting. Default is fill(:lin, length(params))

kwargs : other solver related arguments supported by DiffEqBase.solve. See SciML docs for details

HetaSimulator.gsa — Method

gsa(mcr::MCResults, timepoint::Number)

Computes Pearson Correlation Coeffitients, Partial Regression Coeffietients and Standard Regression Coeffitients for parameters vector and output at a given timepoint

Arguments:

mcr: Monte-Carlo results of type MCResults

timepoint: Time to compute coeffitients at

HetaSimulator.heta_build — Method

heta_build(
  heta_index::AbstractString;
  declaration::String = "platform",
  skip_export::Bool = false,
  log_mode::String = "error",
  debug::Bool = false,
  julia_only::Bool = false,
  dist_dir::String = "dist",
  meta_dir::String = "meta",
  source::String = "index.heta",
  type::String = "heta"
)

Builds the model from Heta-based reactions

See heta comiler docs for details: https://hetalang.github.io/#/heta-compiler/cli-references?id=running-build-with-cli-options

Arguments:

heta_index : path to heta.index file

declaration : path to declaration file. Default is "platform"

skip_export : if set to true no files will be created. Default is false

log_mode : log mode. Default is "error"

debug : turn on debug mode. Default is false

julia_only : export only julia-based model. Default is false

dist_dir : directory path, where to write distributives to. Default is "dist"

meta_dir : meta directory path. Default is "meta"

source : path to the main heta module. Default is "index.heta"

type : type of the source file. Default is "heta"

HetaSimulator.heta_update — Method

heta_update(version::String)

Installs heta-compiler from NPM.

Arguments:

version : heta compiler version. If the value is not provided, heta_update installs the latest version of heta compiler

HetaSimulator.load_jlmodel — Method

load_jlmodel(  
  model_jl::AbstractString; 
  rm_out::Bool = false
)

Loads prebuild julia model without Platform

Arguments:

model_jl : path to Julia model file

rm_out : should the file with Julia model be removed after the model is loaded. Default is false

HetaSimulator.load_jlplatform — Method

load_jlplatform(  
  model_jl::AbstractString; 
  rm_out::Bool = false
)

Loads prebuild julia model as part of Platform

Arguments:

model_jl : path to Julia model file

rm_out : should the file with Julia model be removed after the model is loaded. Default is false

HetaSimulator.load_platform — Method

load_platform(  
  heta_index::AbstractString;
  rm_out::Bool = true,
  julia_only::Bool = true, 
  dist_dir::String = ".",
  source::String = "index.heta",
  type::String = "heta",
  kwargs...
)

Converts heta model to Julia and outputs Platform type.

See heta comiler docs for details: https://hetalang.github.io/#/heta-compiler/cli-references?id=running-build-with-cli-options

Arguments:

heta_index : path to heta.index file

rm_out : should the file with Julia model be removed after the model is loaded. Default is true

julia_only : export only julia-based model. Default is true

dist_dir : directory path, where to write distributives to. Default is "."

kwargs : other arguments supported by heta_build

HetaSimulator.mc — Method

mc(scenario::Scenario,
  params::DataFrame,
  num_iter::Int64;
  kwargs...
)

Run Monte-Carlo simulations with single scenario Scenario. Returns MCResults type.

Example: mc(scn1, DataFrame(k2=rand(3),k3=rand(3)), 1000)

Arguments:

scenario : simulation scenario of type Scenario

params : DataFrame with pre-generated parameters.

num_iter : number of Monte-Carlo iterations

kwargs : other solver related arguments supported by mc(scenario::Scenario, params::Vector, num_iter::Int64)

HetaSimulator.mc — Method

mc(scenario_pairs::Vector{<:Pair},
  params::Vector{<:Pair},
  num_iter::Int64;
  verbose=false,
  alg=DEFAULT_ALG,
  reltol=DEFAULT_SIMULATION_RELTOL,
  abstol=DEFAULT_SIMULATION_ABSTOL,
  parallel_type=EnsembleSerial(),
  kwargs...
)

Run Monte-Carlo simulations with single Scenario. Returns Vector{MCResults} type.

Example: mc([:c1=>scn1,:c2=>scn2], [:k2=>Normal(1e-3,1e-4), :k3=>Uniform(1e-4,1e-2)], 1000)

Arguments:

scenario_pairs : vector of pairs containing names and scenarios of type Scenario

params : parameters variation setup

num_iter : number of Monte-Carlo iterations

verbose : print iteration progress. Default is false

alg : ODE solver. See SciML docs for details. Default is AutoTsit5(Rosenbrock23())

reltol : relative tolerance. Default is 1e-3

abstol : relative tolerance. Default is 1e-6

parallel_type : parallel setup. See SciML docs for details. Default is no parallelism: EnsembleSerial()

kwargs : other solver related arguments supported by DiffEqBase.solve. See SciML docs for details

HetaSimulator.mc — Method

mc(scenario_pairs::Vector{<:AbstractScenario},
  params::Vector{<:Pair},
  num_iter::Int64;
  kwargs...
)

Run Monte-Carlo simulations with single scenario. Returns Vector{MCResults} type.

Example: mc([scn1,scn2], [:k2=>Normal(1e-3,1e-4), :k3=>Uniform(1e-4,1e-2)], 1000)

Arguments:

scenario_pairs : vector of scenarios of type Scenario

params : parameters variation setup

num_iter : number of Monte-Carlo iterations

kwargs : other solver related arguments supported by mc(scenario_pairs::Vector{<:Pair}, params::Vector, num_iter::Int64)

HetaSimulator.mc — Method

mc(model::Model,
  params::Vector{<:Pair},
  num_iter::Int64;
  measurements::Vector{AbstractMeasurementPoint} = AbstractMeasurementPoint[],
  events_active::Union{Nothing, Vector{Pair{Symbol,Bool}}} = Pair{Symbol,Bool}[],
  events_save::Union{Tuple,Vector{Pair{Symbol, Tuple{Bool, Bool}}}}=(true,true), 
  observables::Union{Nothing,Vector{Symbol}} = nothing,
  saveat::Union{Nothing,AbstractVector} = nothing,
  tspan::Union{Nothing,Tuple} = nothing,
  save_scope::Bool=false,
  time_type::DataType=Float64,
  kwargs...
)

Run Monte-Carlo simulations with Model. Returns MCResults type.

Example: mc(model, [:k2=>Normal(1e-3,1e-4), :k3=>Uniform(1e-4,1e-2)], 1000)

Arguments:

model : model of type Model

params : parameters variation setup

num_iter : number of Monte-Carlo iterations

measurements : Vector of measurements. Default is empty vector

events_active : Vector of Pairs containing events' names and true/false values. Overwrites default model's values. Default is empty vector

events_save : Tuple or Vector{Tuple} marking whether to save solution before and after event. Default is (true,true) for all events

observables : names of output observables. Overwrites default model's values. Default is empty vector

saveat : time points, where solution should be saved. Default nothing values stands for saving solution at timepoints reached by the solver

tspan : time span for the ODE problem

save_scope : should scope be saved together with solution. Default is false

kwargs : other solver related arguments supported by mc(scenario::Scenario, params::Vector, num_iter::Int64)

HetaSimulator.mc — Method

mc(platform::Platform, 
  params::Vector{<:Pair},
  num_iter::Int64;
  kwargs...
)

Run Monte-Carlo simulations with single Scenario. Returns Vector{MCResults} type.

Example: mc(platform, [:k2=>Normal(1e-3,1e-4), :k3=>Uniform(1e-4,1e-2)], 1000)

Arguments:

platform : platform of Platform type

params : parameters variation setup

num_iter : number of Monte-Carlo iterations

kwargs : other solver related arguments supported by mc(scenario_pairs::Vector{<:Pair}, params::Vector, num_iter::Int64)

HetaSimulator.mc — Method

mc(scenario::Scenario,
  params::Vector{<:Pair},
  num_iter::Int64;
  verbose=false,
  alg=DEFAULT_ALG,
  reltol=DEFAULT_SIMULATION_RELTOL,
  abstol=DEFAULT_SIMULATION_ABSTOL,
  parallel_type=EnsembleSerial(),
  kwargs...
)

Run Monte-Carlo simulations with single Scenario. Returns MCResults type.

Example: mc(scenario, [:k2=>Normal(1e-3,1e-4), :k3=>Uniform(1e-4,1e-2)], 1000)

Arguments:

scenario : simulation scenario of type Scenario

params : parameters variation setup

num_iter : number of Monte-Carlo iterations

verbose : print iteration progress. Default is false

alg : ODE solver. See SciML docs for details. Default is AutoTsit5(Rosenbrock23())

reltol : relative tolerance. Default is 1e-3

abstol : relative tolerance. Default is 1e-6

parallel_type : parallel setup. See SciML docs for details. Default is no parallelism: EnsembleSerial()

kwargs : other solver related arguments supported by DiffEqBase.solve. See SciML docs for details

HetaSimulator.read_mcvecs — Method

read_mcvecs(filepath::String)

Read table with pre-generated parameters for Monte-Carlo simulations. Typically used for virtual patients simulations

Arguments:

filepath: path to the file with pre-generated parameters

HetaSimulator.read_measurements — Function

read_measurements(filepath::String, sheet=1; kwargs...)

Reads table file with measurements to DataFrame

Arguments:

filepath : path to table file. Supports ".csv" and ".xlsx" files

sheet : number of sheet in case of ".xlsx" file. Default is 1

kwargs : other arguments supported by CSV.File

HetaSimulator.read_scenarios — Function

read_scenarios(filepath::String, sheet=1; kwargs...)

Reads table file with scenarios to DataFrame

Arguments:

filepath : path to table file. Supports ".csv" and ".xlsx" files

sheet : number of sheet in case of ".xlsx" file. Default is 1

kwargs : other arguments supported by CSV.File

HetaSimulator.save_results — Method

save_results(filepath::String, sim::AbstractResults)

Save results as csv file

Arguments:

filepath: path and name of the file to write to

sim: simulation results of AbstractResults type

HetaSimulator.sim — Method

sim(platform::Platform; 
  scenarios::Union{AbstractVector{Symbol}, Nothing} = nothing,
  kwargs...) where {C<:AbstractScenario}

Simulate scenarios included in platform. Returns Vector{Pair}.

Example: sim(platform)

Arguments:

platform : platform of Platform type

scenarios : Vector containing names of scenarios included in platform. Default value nothing stands for all scenarios in the platform

kwargs... : other kwargs supported by sim(scenario_pairs::Vector{Pair})

HetaSimulator.sim — Method

sim(scenarios::AbstractVector{C}; kwargs...) where {C<:AbstractScenario}

Simulate multiple scenarios. Returns Vector{Pair}.

Example: sim([scn1, scn2, scn3])

Arguments:

scenarios : Vector containing names and scenarios of type Scenario

kwargs : other kwargs supported by sim(scenario_pairs::Vector{Pair})

HetaSimulator.sim — Method

sim(scenario::Scenario; 
  parameters_upd::Union{Nothing, Vector{P}}=nothing,
  alg=DEFAULT_ALG, 
  reltol=DEFAULT_SIMULATION_RELTOL, 
  abstol=DEFAULT_SIMULATION_ABSTOL,
  kwargs...)

Simulate single Scenario. Returns SimResults type.

Example: Scenario(model; tspan = (0., 200.), saveat = [0.0, 150., 250.]) |> sim

Arguments:

scenario : simulation scenario of type Scenario

parameters_upd : constants, which overwrite both Model and Scenario constants. Default is nothing

alg : ODE solver. See SciML docs for details. Default is AutoTsit5(Rosenbrock23())

reltol : relative tolerance. Default is 1e-3

abstol : relative tolerance. Default is 1e-6

kwargs... : other solver related arguments supported by DiffEqBase.solve. See SciML docs for details

HetaSimulator.sim — Method

sim(scenario_pairs::Vector{P}; 
  parameters_upd::Union{Nothing, Vector}=nothing,
  alg=DEFAULT_ALG, 
  reltol=DEFAULT_SIMULATION_RELTOL, 
  abstol=DEFAULT_SIMULATION_ABSTOL,
  parallel_type=EnsembleSerial(),
  kwargs...) where P<:Pair

Simulate multiple scenarios. Returns Vector{Pair}.

Example: sim([:x => scn1, :y=>scn2, :z=>scn3])

Arguments:

scenario_pairs : vector of pairs containing names and scenarios of type Scenario

parameters_upd : constants, which overwrite both Model and Scenario constants. Default is nothing

alg : ODE solver. See SciML docs for details. Default is AutoTsit5(Rosenbrock23())

reltol : relative tolerance. Default is 1e-3

abstol : relative tolerance. Default is 1e-6

parallel_type : type of multiple simulations parallelism. Default is no parallelism. See SciML docs for details

kwargs... : other solver related arguments supported by DiffEqBase.solve. See SciML docs for details